+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## 13 Jun 2022
+
+- `changed` compression for downloaded batches from gztar to zip
+
 ## Version `0.26.14` - 9 Jun 2022
 
 - `changed` schemas bump for Microbiome support